Airflow/requirements.txt
pandas
snowflake-connector-python
snowflake-sqlalchemy
pyodbc
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-snowflake
astro-sdk-python 